test update again

+++ app/src/main/java/hexlet/code/dto/urls/UrlsPage.java
@@ -7,15 +7,16 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 import java.util.List;
-import java.util.Map;
+
 
 @AllArgsConstructor
 @Getter
 
 public class UrlsPage extends BasePage {
     private List<Url> urls;
-    private int pageNumber;
+    private long pageNumber;
     private List<UrlCheck> lastCheck;
     private String conditionNext;
     private String conditionBack;
 }
+
